Add approval step for non-master branches
This adds an approval requirement to the
build-and-publish-docker-image job so that
it does not run automatically on non-master
branches and does run automatically on
master branches. This allows developers to
build and publish a development image if they
want one, but does not build those images by
default since they are not used frequently.