gateway: Do not spawn RPC futures on a CpuPool
The jsonrpc libraries use a Tokio runtime by default (if the latest
version of tokio-core is used). This can be overriden by Parity's
Middleware which takes an optional CpuPool argument to spawn the futures
onto.

This commit changes it so None is passed for that optional argument so
that futures are executed on the Tokio runtime.